refactor: create Pages directory and move Home component to it and update imports
