+++ src/App.tsx
@@ -1,6 +1,6 @@
 import { StrictMode } from 'react';
 import { createRoot } from 'react-dom/client';
-import Home from './Home.tsx';
+import Home from './Pages/Home.tsx';
 import { Provider } from 'react-redux';
 import { store } from './store/useTimerStore.ts';
 import './index.css';
